Notif System - On going
The latest version works ... most of the time. There seems to be a big amount of bugs while connecting to the production system ES.

Other errors are related to the fact that its not robust to all the scenarios presented and how we are launching notif reports when it is not necessary.